webkit-efl/LayoutTests/fast/mediastream/getusermedia-expected.txt

Tests webkitGetUserMedia.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.webkitGetUserMedia({audio:false, video:false}, error, error); threw exception Error: NotSupportedError: DOM Exception 9.
PASS navigator.webkitGetUserMedia({audio:true}, gotStream1, error); did not throw exception.
PASS Stream generated.
PASS stream.audioTracks.length is 1
PASS stream.videoTracks.length is 0
PASS {video:true} generated stream
PASS stream.audioTracks.length is 0
PASS stream.videoTracks.length is 1
PASS {audio:true, video:true} generated stream
PASS stream.audioTracks.length is 1
PASS stream.videoTracks.length is 1
PASS successfullyParsed is true

TEST COMPLETE
